+++ otto/__init__.py
@@ -1,3 +1,4 @@
 from router import Router
+from router import Route
 from publisher import Publisher
 from app import Application