+++ CHANGES.rst
@@ -3,10 +3,17 @@ Changes
 
 In next release...
 
+Features
+########
+
+- The ``Route`` class now provides the ``match`` method.
+
 Backwards incompatibilities
 ###########################
 
-- The ``route`` method was renamed to ``connect``.
+- The ``Publisher.route`` method was renamed to ``connect``. This
+  method now takes a route object. This change was also applied for
+  the ``Router`` class.
 
 1.0 (2009-11-12)
 ----------------